# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water Tracker API docs",
    "description": "This is a documentation of Water Tracker app.\n\nThis project is free web-application for comfortable record keeping of drunk water.\n\nThe main goal is to provide users with instruments for control and visualization of water consumption.\n\nThere are possibilities of registration, authentication and authorization of users. They can calculate daily water volume depending on their gender, body weight and sport activities, add/edit/delete time and amount of water they drank. As well, there are list of records for today and month statistic of reaching daily volume goal.\n\nIt is easy and convenient to use app with an intuitive interface, secured by hashing passwords and using access tokens."
  },
  "servers": [
    {
      "url": "https://water-tracker-06.onrender.com"
    }
  ],
...
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get info about user",
        "operationId": "getUser",
        "description": "Get whole available information about user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get information about user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found user with id 66d2e3a18c42147b871bcdcc!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Add/update user's info",
        "operationId": "patchUser",
        "description": "Update/add one or more fields about user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/patchUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update's information about user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User successfully patched!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "_id",
          "email",
          "gender",
          "waterRate",
          "avatar"
        ],
        "properties": {
          "_id": {
            "description": "User's id",
            "type": "string",
            "example": "66d2e3a18c42147b871bcdcc"
          },
          "name": {
            "description": "User's name, can contain max 32 characters",
            "type": "string",
            "example": "Lora"
          },
          "email": {
            "description": "User's email, must be a valid email",
            "type": "string",
            "example": "lora@gmail.com"
          },
          "gender": {
            "description": "User's gender, must be one of [\"male\", \"female\"]",
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male",
            "default": "female"
          },
          "waterRate": {
            "description": "User's daily water goal in ml, can be from 1 to 15000",
            "type": "number",
            "example": 2100,
            "default": 1500
          },
          "avatar": {
            "description": "URL to user's avatar",
            "type": "string",
            "default": null,
            "example": "https://res.cloudinary.com/doicwbvic/image/upload/v1725634437/gvsagq1bnykokhtrtqr4.jpg"
          }
        }
      },
      "patchUser": {
        "type": "object",
        "properties": {
          "name": {
            "description": "User's name, can contain max 32 characters",
            "type": "string",
            "example": "Lora"
          },
          "email": {
            "description": "User's email, must be a valid email",
            "type": "string",
            "example": "lora@gmail.com"
          },
          "gender": {
            "description": "User's gender",
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male",
            "default": "female"
          },
          "waterRate": {
            "description": "User's daily water goal in ml, can be from 1 to 15000",
            "type": "number",
            "example": 2100,
            "default": 1500
          },
          "password": {
            "description": "User's outdated password, can be min 8 max 64 characters",
            "type": "string",
            "example": "123456789"
          },
          "newPassword": {
            "description": "User's new password, can be min 8 max 64 characters",
            "type": "string",
            "example": "987654321"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "BadRequestError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "enum": [
                        "Name must be a string",
                        "Avatar was not received",
                        "Data was not received",
                        "Date should be in the format YYYY-MM-DDTHH:MM:SS"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "enum": [
                        "Access token expired",
                        "Unauthorized",
                        "Session expired",
                        "Session not found",
                        "Incorrect email or password",
                        "Wrong password",
                        "Auth header should be of type Bearer"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "NotFoundError",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "enum": [
                        "User not found",
                        "Route not found",
                        "Water consumption entry not found!"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "ConflictError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email in use"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}